refactor(bench): split json_simd benchmark into focused components
Breaks down the monolithic json_simd benchmark into three specialized
benchmark suites for better organization and more targeted performance
testing of different JSON SIMD operations.

Changes made:
- Remove large json_simd.rs benchmark file (432 lines)
- Add json_simd_common.rs with shared utilities for file discovery
- Add json_simd_indexing.rs for comprehensive SIMD variant testing
- Add json_simd_cursor.rs for simple cursor performance testing
- Add json_simd_full.rs for full index construction and pattern comparison
- Update Cargo.toml with three new benchmark entries
- Update benchmark registry with detailed descriptions for each suite

The split provides better benchmark organization:
- json_simd_indexing: Tests all SIMD variants (AVX2/SSE/NEON) on files up to 100MB
- json_simd_cursor: Tests cursor-only operations on smaller files (up to 10MB)
- json_simd_full: Tests complete JsonIndex construction and pattern fairness

This refactoring improves benchmark maintainability and allows developers
to run specific performance tests based on their optimization focus area.
