+++ src/bin/succinctly/bench_runner/registry.rs
@@ -155,11 +155,29 @@ pub static BENCHMARKS: &[BenchmarkInfo] = &[
         working_dir: ".",
     },
     BenchmarkInfo {
-        name: "json_simd",
-        description: "JSON SIMD operations",
+        name: "json_simd_indexing",
+        description: "JSON SIMD indexing (AVX2/SSE/NEON/PFSM, up to 100MB)",
         category: BenchmarkCategory::Json,
         bench_type: BenchmarkType::Criterion,
-        criterion_name: Some("json_simd"),
+        criterion_name: Some("json_simd_indexing"),
+        cli_subcommand: None,
+        working_dir: ".",
+    },
+    BenchmarkInfo {
+        name: "json_simd_cursor",
+        description: "JSON simple cursor (SIMD/Scalar, up to 10MB)",
+        category: BenchmarkCategory::Json,
+        bench_type: BenchmarkType::Criterion,
+        criterion_name: Some("json_simd_cursor"),
+        cli_subcommand: None,
+        working_dir: ".",
+    },
+    BenchmarkInfo {
+        name: "json_simd_full",
+        description: "JSON full index + pattern comparison",
+        category: BenchmarkCategory::Json,
+        bench_type: BenchmarkType::Criterion,
+        criterion_name: Some("json_simd_full"),
         cli_subcommand: None,
         working_dir: ".",
     },
